Install up CLI & connect to a new Upbound Cloud hosted control plane
This worked at charm:

    make .env -B && . .env
    make up
    up ctp create christmas-2021-gift

    up ctp list
    NAME                  ID                                     SELF-HOSTED   STATUS
    shipit-15-followup    16e3a96e-6505-4ad8-af96-098d5ba1eba4   false         deleting
    christmas-2021-gift   bd465974-fb45-47a4-b9a5-001db23c89c8   false         ready

    m up_connect

While the previous control plane is still deleting, the new one is ready.

As for the new control plane, I was not expecting the following:
- runs on K8s 1.18 - EOS since 2021.04
- uses Crossplane 1.3.1-up.1 - latest is 1.5.1

I think the latter will be a blocker for Terrajet, which I suspect to
require Crossplane 1.4, even though my instinct was to go with 1.5. I
will experiment with the hosted control plane, even though I am
expecting to need to go with a self-hosted Crossplane 1.5.

In terms of feedback, it would have been great to see in the up cli -
up ctp list - which version of Crossplane the control planes hosted in
Upbound are running. Same for K8s. It would have been even more amazing
to be able to install these with min versions, e.g. 1.21 min. for K8s &
1.5 min. version for Crossplane.

Signed-off-by: Gerhard Lazu <gerhard@lazu.co.uk>